extract ResultManager
