+++ flutter/lib/benchmark/result_manager.dart
@@ -1,0 +1,39 @@
+import 'dart:convert';
+import 'dart:io';
+
+class ResultManager {
+  static const _jsonResultFileName = 'result.json';
+  late final String _jsonResultPath;
+
+  ResultManager(String applicationDirectory) {
+    _jsonResultPath = '$applicationDirectory/$_jsonResultFileName';
+  }
+  
+  Future<void> write(List<Map<String, dynamic>> content) async {
+    final jsonFile = File(_jsonResultPath);
+
+    final jsonEncoder = JsonEncoder.withIndent('  ');
+    var encoded = jsonEncoder.convert(content);
+    // needed to match android behavior
+    encoded = encoded.replaceAll('/', '\\/');
+    await jsonFile.writeAsString(encoded);
+  }
+
+  Future<String> get json async {
+    final file = File(_jsonResultPath);
+
+    if (await file.exists()) {
+      return file.readAsString();
+    }
+
+    return '';
+  }
+
+  Future<void> delete() async {
+    var resultsFile = File(_jsonResultPath);
+    if (await resultsFile.exists()) {
+      await resultsFile.delete();
+    }
+  }
+
+}
